Make get_{include,library}_paths part of the public API.
