+++ snapcraft/internal/common.py
@@ -255,3 +255,37 @@ def format_output_in_columns(elements_list, max_width=MAX_CHARACTERS_WRAP,
         result_output.append(sep.join(candidate_output[i]))
 
     return result_output
+
+
+def get_include_paths(envvar, root, arch_triplet, prepend='-I', sep=' '):
+    paths = [
+        os.path.join(root, 'include'),
+        os.path.join(root, 'usr', 'include'),
+        os.path.join(root, 'include', arch_triplet),
+        os.path.join(root, 'usr', 'include', arch_triplet),
+    ]
+
+    include_paths = ['{}{}'.format(prepend, p)
+                     for p in paths if os.path.exists(p)]
+    return _combine_paths(envvar, include_paths, prepend, sep)
+
+
+def get_library_paths(envvar, root, arch_triplet, prepend='-L', sep=' '):
+    paths = [
+        os.path.join(root, 'lib'),
+        os.path.join(root, 'usr', 'lib'),
+        os.path.join(root, 'lib', arch_triplet),
+        os.path.join(root, 'usr', 'lib', arch_triplet),
+    ]
+
+    library_paths = ['{}{}'.format(prepend, l)
+                     for l in paths if os.path.exists(l)]
+    return _combine_paths(envvar, library_paths, prepend, sep)
+
+
+def _combine_paths(envvar, paths, prepend, sep):
+    if not paths:
+        return ''
+    else:
+        return '{envvar}="${envvar}{sep}{paths}"'.format(
+            envvar=envvar, sep=sep, paths=sep.join(paths))
